requirements.txt
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
web3==6.20.2
